Created 2DCoordinates header
