+++ src/chunk.h
@@ -1,40 +1,9 @@
 #pragma once
 
 #include "./section.h"
+#include <2DCoordinates.hpp>
 #include <map.hpp>
 #include <nbt/nbt.hpp>
-
-struct Coordinates {
-  int32_t x, z;
-
-  template <typename I, std::enable_if_t<std::is_integral<I>::value, int> = 0>
-  Coordinates(std::initializer_list<I> l) {
-    x = *l.begin();
-    z = *(l.begin() + 1);
-  }
-
-  Coordinates operator+(const Coordinates &other) const {
-    return {x + other.x, z + other.z};
-  }
-
-  bool operator<(const Coordinates &other) const {
-    // Code from <bits/stl_pair.h>
-    return x < other.x || (!(other.x < x) && z < other.z);
-  }
-
-  bool operator==(const Coordinates &other) const {
-    return x == other.x && z == other.z;
-  }
-
-  bool operator!=(const Coordinates &other) const {
-    return !this->operator==(other);
-  }
-
-  template <typename I, std::enable_if_t<std::is_integral<I>::value, int> = 0>
-  Coordinates operator%(const I &m) const {
-    return {x % m, z % m};
-  }
-};
 
 namespace mcmap {
